Reimplement most data fetching
We'll now use the XML API, because of its support for XPath queries.
This allows us to be more precise about what data to fetch and when.
One upside is we use about 50% less data now. Another is that because
we no longer have to blindly fetch all builds, we can display version
tables one-by-one. We can also implement per-version refreshing.